refactor(type-checker): extract ProofChecks partial class [slice 4/6]
Move 9 proof-analysis methods from PreceptTypeChecker.cs into
PreceptTypeChecker.ProofChecks.cs — pure structural refactor, no
behavior changes.

Methods moved:
- ExtractFieldInterval
- TryInferInterval
- FlipComparisonOperator
- AssessDivisorSafety
- AssessGuard
- TryExtractSingleFieldComparison
- DescribeExpression
- AssessNonnegativeArgument
- TryGetNumericLiteral

FlipComparisonOperator is called cross-file from Narrowing
(TryApplyNumericComparisonNarrowing) — intentional, noted in plan.

Closes #118 (partial — slice 4 of 6)
Ref: PR #123
